Backend/requirements.txt
Flask
Flask-Cors
SQLAlchemy
PyMySQL
python-dotenv
chromadb
langchain
langchain-google-genai
langchain-community
pypdf
